// frontend/src/types/registerRequest.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface RegisterRequest {
  /** @minLength 1 */
  verificationToken: string;
  email?: string;
  /** @minLength 1 */
  fullName: string;
  /** @minLength 1 */
  password: string;
}